# agent-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Weavster API
  version: 0.1.0
  description: >
    REST API for the Weavster message-oriented integration platform.
    JSON-first; the read-only topology graph is the confirmed MVP UI surface.
paths:
  /api/v1/auth/login:
    post:
      summary: Log in and receive a bearer token (valid 12 hours)
      security:
        - csrfMarker: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string}
                mfaCode: {type: string}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: "Token, expiresAt, and user"}
        "401": {description: Invalid username or password}
  /api/v1/auth/logout:
    post:
      summary: Revoke the bearer token used for this request
      responses:
        default: {$ref: "#/components/responses/Error"}
        "204": {description: Logged out}
  /api/v1/auth/me:
    get:
      summary: Current user, permissions, and whether a password change is required
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: Current user}
  /api/v1/auth/password:
    post:
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: {type: string}
                newPassword: {type: string}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "204": {description: Password changed}
        "400": {description: New password rejected by the password policy}
  /api/v1/system:
    get:
      summary: System status
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: System information (id, version, build date, runtime, charsets, protocols/ciphers, license).
  /api/v1/topology:
    get:
      summary: Flow topology overview graph (read-only)
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: Overview graph (flow nodes + route/dependency edges).
  /api/v1/topology/flows/{flowId}:
    get:
      summary: Flow-internal topology graph (read-only)
      parameters:
        - name: flowId
          in: path
          required: true
          schema:
            type: string
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: Flow-internal graph (source -> transform -> destination nodes + message-path edges).
  /api/v1/flows:
    get:
      summary: List flows
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: Flow list.
    post:
      summary: Create flow
      responses:
        default: {$ref: "#/components/responses/Error"}
        "201":
          description: Created.
    put:
      summary: Replace several flow definitions; each keeps its status, stopped destinations, and (unless set) enabled (requires flows:edit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "{flows: [flow definitions]}; no other top-level field"
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: "{updated: [ids]}"
        "400":
          description: Invalid document, flow, or dependencies; nothing written.
        "404":
          description: Unknown flows (listed); nothing written.
        "413":
          description: Document larger than 50 MiB.
        "500":
          description: "Stopped part-way: {error: {code: UPDATE_INCOMPLETE}, updated: [ids]} lists what was written."
  /api/v1/flows/connector-names:
    get:
      summary: List every flow's source type and destination names (requires flows:view)
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: "[{id, name, sourceType, destinations: [names]}]"
  /api/v1/flows/ports-in-use:
    get:
      summary: List the ports the server listens on (requires flows:view)
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: "[{address, port, usedBy}]"
  /api/v1/flows/{id}:
    put:
      summary: Replace a flow's definition; status is kept (requires flows:edit)
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: The updated flow}
        "400": {description: Invalid definition, a status field, or a different id in the body}
        "404": {description: Unknown flow}
    get:
      summary: Get flow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: Flow record.
    delete:
      summary: Delete flow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        default: {$ref: "#/components/responses/Error"}
        "204":
          description: Deleted.
        "404":
          description: Unknown flow.
        "409":
          description: Other flows depend on this flow.
  /api/v1/flows/{id}/enable:
    post:
      summary: Make a flow eligible for auto-deploy at startup (requires flows:edit)
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: The updated flow}
        "404": {description: Unknown flow}
  /api/v1/flows/{id}/disable:
    post:
      summary: Remove a flow's auto-deploy eligibility (requires flows:edit)
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: The updated flow}
        "404": {description: Unknown flow}
  /api/v1/flows/{id}/destinations/{name}/{action}:
    post:
      summary: Stop (hold deliveries to) or start one destination of a flow (requires flows:deploy)
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: name, in: path, required: true, schema: {type: string}}
        - {name: action, in: path, required: true, schema: {type: string, enum: [start, stop]}}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: "The flow; stoppedDestinations lists stopped destinations"}
        "404": {description: Unknown flow, destination, or action}
  /api/v1/flows/{id}/{action}:
    post:
      summary: Change a flow's lifecycle state (requires flows:deploy)
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
        - name: action
          in: path
          required: true
          schema: {type: string, enum: [deploy, undeploy, start, stop, pause, halt, resume]}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: The updated flow}
        "404": {description: Unknown flow or action}
        "409": {description: "Transition not allowed from the flow's status"}
  /api/v1/flows/export:
    get:
      summary: Export flow definitions plus their transitive dependencies (requires flows:view)
      parameters:
        - name: ids
          in: query
          description: Comma-separated flow ids; omit for all flows
          schema: {type: string}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: "{version: 1, flows: [definitions without status]}"}
        "404": {description: Unknown flow id}
  /api/v1/flows/import:
    post:
      summary: Import an export document, validated as a whole before writing (requires flows:edit)
      parameters:
        - name: overwrite
          in: query
          schema: {type: boolean}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: "{created: [ids], updated: [ids]}"}
        "400": {description: Invalid document, flow, or dependencies; nothing written}
        "409": {description: Flows already exist and overwrite is not set; nothing written}
        "413": {description: Document larger than 50 MiB}
        "500": {description: "Stopped part-way: {error: {code: IMPORT_INCOMPLETE}, created: [ids], updated: [ids]} lists what was written"}
  /api/v1/flows/redeploy-all:
    post:
      summary: Undeploy and re-deploy every flow that is not undeployed (deployed, started, paused, halted, stopped); each ends deployed (requires flows:deploy)
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: The redeployed flows}
        "500": {description: "Stopped part-way: {error: {code: REDEPLOY_INCOMPLETE}, redeployed: [flows already redeployed]}"}
  /api/v1/flows/{id}/messages:
    post:
      summary: Send a message into a flow (requires messages:send)
      description: >
        Persists the message, runs the flow's transform, and delivers it to each destination.
        The body must be a single JSON object when the flow has a transform; passthrough
        flows (no transform) accept any bytes.
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
          application/octet-stream:
            schema: {type: string, format: binary}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "202": {description: "Processed: {id, status, response?} with status sent, queued, dead-lettered, filtered, or errored; response is the responseSelector destination's reply"}
        "400": {description: The flow or a destination has a transform (a responseTransform does not count) and the body is not a single JSON object, or the body could not be read}
        "404": {description: Unknown flow}
        "409": {description: The flow is not started}
        "413": {description: Body larger than 10 MiB}
  /api/v1/flows/{id}/stats:
    get:
      summary: Flow message counters (requires flows:view)
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
        - name: lifetime
          in: query
          schema: {type: boolean}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: "received, filtered, transformed, sent, errored, queued, destinations, lastMessageAt"}
        "400": {description: lifetime is not a boolean}
        "404": {description: Unknown flow}
  /api/v1/events:
    get:
      summary: Search the event log (requires events:view)
      parameters:
        - name: type
          in: query
          schema: {type: string}
        - name: flowId
          in: query
          schema: {type: string}
        - name: limit
          in: query
          description: Newest N matching events (1-10000, default 1000)
          schema: {type: integer, minimum: 1, maximum: 10000}
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200": {description: "Events, oldest first: id, at, type, flowId, data"}
        "400": {description: Invalid limit}
  /api/v1/messages:
    get:
      summary: Search messages
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: flowId
          in: query
          schema:
            type: string
      responses:
        default: {$ref: "#/components/responses/Error"}
        "200":
          description: Message search results.
security:
  - basicAuth: []
    csrfMarker: []
  - bearerAuth: []
    csrfMarker: []
components:
  securitySchemes:
    basicAuth: {type: http, scheme: basic}
    bearerAuth: {type: http, scheme: bearer}
    csrfMarker: {type: apiKey, in: header, name: X-Weavster-CSRF}
  schemas:
    Error:
      type: object
      description: "Every error reply (see docs/api-errors.md)."
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, description: "Stable code, e.g. NOT_FOUND, CONFLICT, NOT_IMPLEMENTED"}
            message: {type: string}
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}